requirements.txt
boto3==1.40.56
botocore==1.40.56

strands-agents==1.13.0
strands-agents-tools==0.2.12
strands-agents-builder==0.1.10

uv

fastapi[standard]==0.119.1
pydantic==2.12.3

mcp==1.18.0
fastmcp==2.12.4

ddgs
